refactor: modularize advisory policy live proof
