+++ tests/unit/live-canonical-validation-script.test.ts
@@ -201,6 +201,14 @@ describe("canonical live validation script", () => {
       join(process.cwd(), "scripts", "live", "validation", "browser-workflows.mjs"),
       "utf8"
     );
+    const advisoryPolicyProof = readFileSync(
+      join(process.cwd(), "scripts", "live", "validation", "advisory-policy-proof.mjs"),
+      "utf8"
+    );
+    const payloadUtils = readFileSync(
+      join(process.cwd(), "scripts", "live", "validation", "payload-utils.mjs"),
+      "utf8"
+    );
 
     expect(script).toContain("assertPerformanceCalculationSanity");
     expect(script).toContain("assertRiskCalculationSanity");
@@ -215,16 +223,18 @@ describe("canonical live validation script", () => {
     expect(script).toContain("validateAdvisoryJourneyScreens");
     expect(script).toContain("validateProposalNarrativePosturePanel");
     expect(script).toContain("validateProposalMemoEvidencePackPanel");
+    expect(script).toContain('from "./validation/advisory-policy-proof.mjs"');
     expect(script).toContain("Create proposal narrative canonical proof");
-    expect(script).toContain("Create advisory policy evaluation canonical proof");
-    expect(script).toContain("Advisory policy review queue canonical proof");
-    expect(script).toContain("POLICY_EVALUATION_PENDING_REVIEW_CREATED");
-    expect(script).toContain("POLICY_PACK_VERSION_ALREADY_ACTIVE_IMMUTABLE");
-    expect(script).toContain('status: "already_active"');
-    expect(script).toContain('import { createHash } from "node:crypto"');
+    expect(advisoryPolicyProof).toContain("Create advisory policy evaluation canonical proof");
+    expect(advisoryPolicyProof).toContain("Advisory policy review queue canonical proof");
+    expect(advisoryPolicyProof).toContain("POLICY_EVALUATION_PENDING_REVIEW_CREATED");
+    expect(advisoryPolicyProof).toContain("POLICY_PACK_VERSION_ALREADY_ACTIVE_IMMUTABLE");
+    expect(advisoryPolicyProof).toContain('status: "already_active"');
+    expect(payloadUtils).toContain('import { createHash } from "node:crypto"');
+    expect(payloadUtils).toContain("buildPayloadScopedIdempotencyKey");
     expect(script).toContain("buildPayloadScopedIdempotencyKey");
     expect(script).toContain("proposalCreateIdempotencyKey");
-    expect(script).toContain("wb-policy-evaluation");
+    expect(advisoryPolicyProof).toContain("wb-policy-evaluation");
     expect(script).toContain("proofPackIdempotencyKey");
     expect(script).toContain("proposal.narrative_posture");
     expect(script).toContain("proposal.memo_evidence_pack");
